Force StorageAccount name to Lowercase / DnsServer IP also needs to be set on NetworkInterface
StorageAccount:
Just to avoid mistakes while deploying this template. When you enter a StorageAccount-Name with an Uppercase character it fails. The "toLower" feature of ARM fixes this.
Future improvement: Do the same via variables as  the "newStorageAccountName"-parameter is still used in various other places and might have an Uppercase character. But that does not hurt for now.

DNSServers:
For the VirtualNetwork, the IP address of the DNS Server is hardcoded. However, it seems like it does not propagate through to the NetworkInterface (don't know if it should do that automatically). Anyway, the same IP is hardcoded for the NetworkInterface now.