Update template with the following changes
1) Keeping the template naming to azuredepoly.parameters.json for the
parameter file. We will keep the original JSON as azuredeploy.json. If
we need to keep a version of this for Azure compatibility we will create
a checkin into Azure quick start template so there is consistency
between Azurestack and Azure templates everywhere.
2) Update the parameter schema to deploymentParameters.json
3) Reduce the number of parameters. We will hard code the best possible
size for the VM, Subnet, storageType, assetLocation , IPAddressRange and
IPAddress. We will also keep the location of all the resources as the
resource group. We will update the readme to reflect this and let folks
change it if they need to.
4) Ran this though JSON online formatter (will check and update if the
VS native formatter is same as this)
5) Moved all the APIVersions to a variable.
6) Updated all the resource names with the resource group name. Will
also update this in the readme to ensure that RG name is not given
something which can potentially fail.
7) Add output even if it is blank to all the files
8) For every resource group tried to keep the flow as name, type,
location, apiversion, dependson and properties so that it looks uniform
all over.